services/messenger/requirements.txt
python-telegram-bot==20.8
redis
python-dotenv
